Move base derivation to config
